File tree working
File tree is working 100%, still some minor tweaking to be done but all is functional
